admin y permisos a rutas privadas

+++ src/routes/restaurantsRoutes/restaurantsHandlers/deleteHandler.js
@@ -3,8 +3,12 @@ const deleteRestaurantController = require("../restaurantsControllers/deleteRest
 const handlerDeleteData = async (req, res) => {
     try {
         const { id } = req.params
-        const restaurant = await deleteRestaurantController(id);
-        res.status(200).json({ restaurant });
+        const {_id, isAdmin} = req.user;
+        if(id==_id || isAdmin){
+            const restaurant = await deleteRestaurantController(id);
+            res.status(200).json({ restaurant });
+        }
+        else{throw new Error("You can't modify this restaurant")}
     } catch (error) {
         console.log(error);
     }
